Enhanced the output reduction
Created new output scenarious in the config.json file
for NaN check, struct fields check, deadlock check and data races check
Created the reduce_output field in the config.json file to tell the
LLM to produce a smaller output
Created new regex patterns in the reduce_output2() function
Created the remove_patterns_nltk that uses NLP methods to identify
patterns in a sequence of tokenized words. The advantage of it are
that it provides more granular control over text sentences, it can
be more precise in some situations.
Created the test_output_reducer file that tests the functionality of the esbmc_output_optimisation(),
reduce_output2() and remove_patterns_nltk() functions that covers all the string patterns
Instantiated the GPT_4_TURBO_PREVIEW model in the enum of the ai_models.py and used it for various
code bases.

Signed-off-by: mihai.state <mihaita.state@yahoo.com>