Modularized ESBMC-AI further:
- UserChatCommand now handles the user chat mode,
it is the default command that is invoked when the
command runner is not activated.
- Verifier is not called by default, it is up to
the command to handle that.
- Finalized strategy for resolving addon config
fields. We now use inspection to copy all values
automatically. Name is the field that is resolved.
- BaseConfig now provides additional details when
field validation fails.
- ConfigField: Added get_config_message lambda
function.
- Config: Added field for loading multiple files
from the config and args.
- Config: Added field for specifying a default
entry function.
- Solution: Added option for loading and creating
source files.
- FixCodeCommand: Now calls the verifier at the
beginning.
- ESBMC: Added option to specify default entry
function.